protect delete, upddate organization

+++ packages/mcp-tools-logto/src/tools/deleteOrganization.ts
@@ -1,4 +1,4 @@
-import { Tool } from "@coeus-agent/mcp-tools-base";
+import { AuthInfo, Tool } from "@coeus-agent/mcp-tools-base";
 import { ToolCallback } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { z, ZodRawShape } from "zod";
 
@@ -9,8 +9,26 @@ const inputSchema = {
 };
 
 function getCallback(client: LogToClient): ToolCallback<typeof inputSchema> {
-    return async (params) => {
+    return async (params, { authInfo }) => {
         const { id } = params;
+        const { subject } = authInfo! as AuthInfo;
+
+        const roles = (await client.GET("/api/organizations/{id}/users/{userId}/roles", {
+            params: {
+                path: {
+                    id,
+                    userId: subject!,
+                },
+            },
+        })).data!;
+
+        if (roles.some((r: { name: string }) => r.name === "owner") === false) {
+            return {
+                content: [
+                    { type: "text", text: JSON.stringify({ error: "User is not authorized to delete this organization." }) },
+                ],
+            };
+        }
 
         const response = await client.DELETE("/api/organizations/{id}", {
             params: {
